// src/types/types.ts
export type User = {
  avatarUrl: string;
  bio: string | null;
  blog: string | null;
  company: string | null;
  createdAt: Date;
  email: string | null;
  eventsUrl: string;
  followers: number;
  followersUrl: string;
  following: number;
  followingUrl: string;
  gistsUrl: string;
  gravatarId: string;
  htmlUrl: string;
  id: number;
  location: string;
  login: string;
  name: string;
  nodeId: string;
  organizationsUrl: string;
  publicGists: string;
  publicRepos: string;
  receivedEventsUrl: string;
  reposUrl: string;
  siteAdmin: boolean;
  starredUrl: string;
  subscriptionsUrl: string;
  twitterUsername: string;
  type: string;
  updatedAt: string;
  url: string;
};